Game state race conditions
Modifying the game state is full of race conditions (think of ships as threads) and leaving the game state in an inconsistent state (people may be playing with require commnet for control = false, when they expect = true, and have set it to true).

Did some testing, this looks like it works as expected.

Will be able to control a near by ship with the probe control port as well (tested both multi hop and single hop probe controls). (My test case had a RA-100 relay antenna on it, in case that's needed to control near by ships).