Implement IRelayEnabler for AI mode
"""Any module that implements this interface can make all antennae, not just those of type RELAY, work as a relay."""

From the https://kerbalspaceprogram.com/api/class_comm_net_1_1_module_relay_enabler.html

This would allow the XFELTs to be used as an interstellar relay antenna, and gives another use for the AI mode. This will work for stock commnet, and commnet constellation mod by the looks of it.