Initial support for Real Antennas
Fortunately this is easy as we have the dish / panel size available to us.

Supporting Real Antennas is necessary if the user is also using Kerbalism, as Real Antennas uses the Kerbalism API so that it can determine if ships are contactable. Without these changes, our antennas won't work with Kerbalism's background processing.